requirements.txt
fastapi
uvicorn
python-dotenv
SQLAlchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
requests
python-multipart